+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## **2.0.2** - *2020-xx-xx*
-* Gallery - `GALLERY_NEW_PHOTOS_FIRST` option added;
+* Gallery:
+    * `GALLERY_NEW_PHOTOS_FIRST` option added;
+    * resize photoset cover;
 * API - `REGISTRATION_MODE` option added (possible values are `open` and `invite`);
 
 ## **2.0.1** - *2020-04-01*